initial commit
This classification has been made on a Medical-Mnist dataset that contains some scanned images of 6 different categories belonging to the human body’s internal structure. The methodology carries the classification task using Convolutional Neural Network, carefully designed for the classification of images according to the classes to which they belong. The Convolutional Neural Network is stacked with different layers; Input, hidden, and output, and also been optimized with the help of wavelet transform image preprocessing technique and l1 regularization method so as to improve the network architecture. The network has been designed with the strategy of building a network which is less overfitted along with optimized performance.